+++ scripts/test/Selenium/Agent/Admin/AdminQueueTemplates.t
@@ -37,10 +37,12 @@ $Selenium->RunTest(
             UserLogin => $TestUserLogin,
         );
 
-        # create test queue
-        my $QueueRandomID = "queue" . $Helper->GetRandomID();
-        my $QueueID       = $Kernel::OM->Get('Kernel::System::Queue')->QueueAdd(
-            Name            => $QueueRandomID,
+        my $ScriptAlias = $Kernel::OM->Get('Kernel::Config')->Get('ScriptAlias');
+
+        # add test queue
+        my $QueueName = "queue" . $Helper->GetRandomID();
+        my $QueueID   = $Kernel::OM->Get('Kernel::System::Queue')->QueueAdd(
+            Name            => $QueueName,
             ValidID         => 1,
             GroupID         => 1,
             SystemAddressID => 1,
@@ -51,31 +53,38 @@ $Selenium->RunTest(
         );
         $Self->True(
             $QueueID,
-            "Created Queue - $QueueRandomID",
+            "Created Queue - $QueueName",
         );
 
         # get standard template object
         my $StandardTemplateObject = $Kernel::OM->Get('Kernel::System::StandardTemplate');
+        my @Templates;
 
         # create test template
-        my $TemplateRandomID = "template" . $Helper->GetRandomID();
-        my $TemplateID       = $StandardTemplateObject->StandardTemplateAdd(
-            Name         => $TemplateRandomID,
-            Template     => 'Thank you for your email.',
-            ContentType  => 'text/plain; charset=utf-8',
-            TemplateType => 'Answer',
-            ValidID      => 1,
-            UserID       => $UserID,
-        );
-        $Self->True(
-            $QueueID,
-            "Created StandardTemplate - $TemplateRandomID",
-        );
+        for ( 1 .. 2 ) {
+            my $StandardTemplateName = "standard template" . $Helper->GetRandomID();
+            my $TemplateID           = $StandardTemplateObject->StandardTemplateAdd(
+                Name         => $StandardTemplateName,
+                Template     => 'Thank you for your email.',
+                ContentType  => 'text/plain; charset=utf-8',
+                TemplateType => 'Answer',
+                ValidID      => 1,
+                UserID       => $UserID,
+            );
 
-        # get script alias
-        my $ScriptAlias = $Kernel::OM->Get('Kernel::Config')->Get('ScriptAlias');
+            $Self->True(
+                $TemplateID,
+                "Test StandardTemplate is created - $TemplateID"
+            );
 
-        # navigate to AdminQueueTemplates screen
+            my %Template = (
+                TemplateID => $TemplateID,
+                Name       => $StandardTemplateName,
+            );
+            push @Templates, \%Template;
+        }
+
+        # check overview AdminQueueTemplates screen
         $Selenium->VerifiedGet("${ScriptAlias}index.pl?Action=AdminQueueTemplates");
 
         # check overview AdminQueueTemplates
@@ -90,41 +99,74 @@ $Selenium->RunTest(
 
         # check for test template and test queue on screen
         $Self->True(
-            index( $Selenium->get_page_source(), $TemplateRandomID ) > -1,
-            "$TemplateRandomID found on screen"
+            index( $Selenium->get_page_source(), $Templates[0]->{Name} ) > -1,
+            "$Templates[0]->{Name} found on screen"
         );
         $Self->True(
-            index( $Selenium->get_page_source(), $QueueRandomID ) > -1,
-            "$QueueRandomID found on screen"
+            index( $Selenium->get_page_source(), $QueueName ) > -1,
+            "$QueueName found on screen"
         );
 
         # test search filters
-        $Selenium->find_element( "#FilterTemplates", 'css' )->send_keys($TemplateRandomID);
-        $Selenium->find_element( "#FilterQueues",    'css' )->send_keys($QueueRandomID);
+        $Selenium->find_element( "#FilterTemplates", 'css' )->send_keys( $Templates[0]->{Name} );
+        $Selenium->find_element( "#FilterQueues",    'css' )->send_keys($QueueName);
         sleep 1;
 
         $Self->True(
-            $Selenium->find_element("//a[contains(\@href, \'Subaction=Template;ID=$TemplateID' )]")->is_displayed(),
-            "$TemplateRandomID found on screen with filter on",
+            $Selenium->find_element("//a[contains(\@href, \'Subaction=Template;ID=$Templates[0]->{TemplateID}' )]")
+                ->is_displayed(),
+            "$Templates[0]->{Name} found on screen with filter on",
         );
 
         $Self->True(
             $Selenium->find_element("//a[contains(\@href, \'Subaction=Queue;ID=$QueueID' )]")->is_displayed(),
-            "$QueueRandomID found on screen with filter on",
+            "$QueueName found on screen with filter on",
         );
 
-        # change test Queue relation for test Queue
-        $Selenium->find_element("//a[contains(\@href, \'Subaction=Template;ID=$TemplateID' )]")->VerifiedClick();
-
+        # change test Queue relation for the first Template
+        $Selenium->find_element("//a[contains(\@href, \'Subaction=Template;ID=$Templates[0]->{TemplateID}' )]")
+            ->click();
         $Selenium->find_element("//input[\@value='$QueueID'][\@type='checkbox']")->click();
         $Selenium->find_element("//button[\@value='Save'][\@type='submit']")->VerifiedClick();
 
-        # check test Template relation for test Queue
+        # change test Template relation for test Queue
         $Selenium->find_element("//a[contains(\@href, \'Subaction=Queue;ID=$QueueID' )]")->VerifiedClick();
+        $Selenium->find_element("//input[\@value='$Templates[1]->{TemplateID}'][\@type='checkbox']")->click();
 
+        # test checked and unchecked values while filter is used for Template
+        # test filter with "WrongFilterTemplate" to uncheck all values
+        $Selenium->find_element( "#Filter", 'css' )->clear();
+        $Selenium->find_element( "#Filter", 'css' )->send_keys("WrongFilterTemplate");
+        sleep 1;
+
+        # test is no data matches
         $Self->True(
-            $Selenium->find_element("//input[\@value='$TemplateID'][\@type='checkbox']")->is_selected(),
-            "$QueueRandomID is in a relation with $TemplateRandomID",
+            $Selenium->find_element( ".FilterMessage.Hidden>td", 'css' )->is_displayed(),
+            "'No data matches' is displayed'"
+        );
+
+        # check template filter with existing Template
+        $Selenium->find_element( "#Filter", 'css' )->clear();
+        $Selenium->find_element( "#Filter", 'css' )->send_keys( $Templates[1]->{Name} );
+        sleep 1;
+
+        # uncheck the second test standard template
+        $Selenium->find_element("//input[\@value='$Templates[1]->{TemplateID}'][\@type='checkbox']")->click();
+
+        # test checked and unchecked values after using filter
+        $Selenium->find_element( "#Filter", 'css' )->clear();
+        $Selenium->find_element( "#Filter", 'css' )->send_keys("StandardTemplate");
+        sleep 1;
+
+        $Self->Is(
+            $Selenium->find_element("//input[\@value='$Templates[0]->{TemplateID}'][\@type='checkbox']")->is_selected(),
+            1,
+            "$QueueName is in a relation with $Templates[0]->{Name}",
+        );
+        $Self->Is(
+            $Selenium->find_element("//input[\@value='$Templates[1]->{TemplateID}'][\@type='checkbox']")->is_selected(),
+            0,
+            "$QueueName is not in a relation with $Templates[1]->{Name}",
         );
 
         # since there are no tickets that rely on our test QueueTemplate,
@@ -145,13 +187,18 @@ $Selenium->RunTest(
             );
             $Self->True(
                 $Success,
-                "Deleted - $QueueRandomID",
+                "Deleted queue- $QueueName",
             );
         }
 
-        if ($TemplateID) {
+        for my $Template (@Templates) {
             $Success = $StandardTemplateObject->StandardTemplateDelete(
-                ID => $TemplateID,
+                ID => $Template->{TemplateID},
+            );
+
+            $Self->True(
+                $Success,
+                "Deleted StandardTemplate - $Template->{TemplateID}",
             );
         }